move canvas code
